FEAT - Windows packaging via Docker - added new Dockerfile, `Dockerfile-Wine.build` that creates a Wine environment for Windows builds; - changes to `buildwin.sh` to persuade it to run headless under wine and Docker:   * started refactoring versions for `WinPython` downloads but still needs a little work;   * disable certificate checking for `WinPython` downloads from SourceForge;   * now simply unzipping the `WinPython` download to the appropriate `~/.wine` directory as this works headless; and   * tidied `shellcheck` lint errors; - using `docker-package-win.sh` script to build windows installers for now; and - testing new script under Travis.

+++ buildwin.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 
-# Build 64 and 32 bit Windows binaries using Wine and WinPython. If the required Wine environment 
+# Build 64 and 32 bit Windows binaries using Wine and WinPython. If the required Wine environment
 # (WinPython + PyInstaller) cannot be found it is set up quasi-automatically (the WinPython installer
 # needs some manual input)
-# Precondition: 64-bit version of Wine is already installed. 
-
+# Precondition: 64-bit version of Wine is already installed.
+export DISPLAY=:0.0
 # WinPython download URLS
-downloadURL64bit=https://sourceforge.net/projects/winpython/files/WinPython_2.7/2.7.13.1/WinPython-64bit-2.7.13.1Zero.exe/download
-downloadURL32bit=https://sourceforge.net/projects/winpython/files/WinPython_2.7/2.7.13.1/WinPython-32bit-2.7.13.1Zero.exe/download
-
+winPython2Major=2.7
+winPython3Major=3.5
+winPython2Ver="${winPython2Major}.13.1"
+winPython3Ver="${winPython3Major}.3.1"
+downloadURLP264bit="https://sourceforge.net/projects/winpython/files/WinPython_${winPython2Major}/${winPython2Ver}/WinPython-64bit-${winPython2Ver}Zero.exe/download"
+downloadURLP232bit="https://sourceforge.net/projects/winpython/files/WinPython_${winPython2Major}/${winPython2Ver}/WinPython-32bit-${winPython2Ver}Zero.exe/download"
+downloadURLP364bit="https://sourceforge.net/projects/winpython/files/WinPython_${winPython3Major}/${winPython3Ver}/WinPython-64bit-${winPython3Ver}Zero.exe/download"
+downloadURLP332bit="https://sourceforge.net/projects/winpython/files/WinPython_${winPython3Major}/${winPython3Ver}/WinPython-32bit-${winPython3Ver}Zero.exe/download"
 # PyInstaller spec files that defines build options
 specFile64bit=jpylyzer_win64.spec
 specFile32bit=jpylyzer_win32.spec
@@ -25,14 +30,14 @@ installPython(){
     # - $1: bitness (32 or 64)
     # - $2: download URL
     echo "Downloading installer"
-    wget $2 -O pyTemp.exe
+    wget $2 -q --no-check-certificate -O pyTemp.exe
     echo ""
     echo "Installing Python. This requires some user input"
     echo ""
-    echo "Follow installer instructions. For Destination Folder replace default path with C:\Python27_"$1
+    echo "Follow installer instructions. For Destination Folder replace default path with C:\Python27_${1}"
     echo ""
 
-    WINEDEBUG=$WineDebug wine pyTemp.exe
+    7z x -o"$HOME/.wine/drive_c/Python27_$1" pyTemp.exe
 
     echo "Removing installer"
     rm pyTemp.exe
@@ -41,7 +46,7 @@ installPython(){
 installPyInstaller(){
     # Installs pyInstaller if it is not installed already. Argument:
     # - $1: python (full path of python interpreter)
-    echo "Checking for pyinstaller" 
+    echo "Checking for pyinstaller ${1}"
     WINEDEBUG=$WineDebug wine $1 -m pip show pyinstaller
 
     if [ $? -eq 0 ]; then
@@ -53,12 +58,12 @@ installPyInstaller(){
 }
 
 buildBinaries(){
-    # Builds Windows binaries. 
-    
+    # Builds Windows binaries.
+
     # Read arguments:
     bitness=$1
     pyRoot=$2
-    pyInstallerWine=$pyRoot"/Scripts/pyinstaller.exe"
+    pyInstallerWine="${pyRoot}/Scripts/pyinstaller.exe"
     pythonWine=$3
     specFile=$4
 
@@ -66,11 +71,11 @@ buildBinaries(){
     workDir=$PWD
 
     # Directory where build is created (should be identical to 'name' in 'coll' in spec file!!)
-    distDir=$workDir"/dist/win"$bitness"/"
+    distDir="${workDir}/dist/win${bitness}/"
 
-    # Executes jpylyzer with -v option and stores output to 
+    # Executes jpylyzer with -v option and stores output to
     # env variable 'version'
-    # Also trim trailing EOL character and replace '.' by '_' 
+    # Also trim trailing EOL character and replace '.' by '_'
     WINEDEBUG=$WineDebug wine $pythonWine -m $scriptBaseName -v 2> temp.txt
     version=$(head -n 1 temp.txt | tr -d '\r')
     rm temp.txt
@@ -79,53 +84,53 @@ buildBinaries(){
     WINEDEBUG=$WineDebug wine $pyInstallerWine $specFile --distpath=$distDir
 
     # Generate name for ZIP file
-    zipName=$scriptBaseName"_"$version"_win"$bitness".zip"
+    zipName="${scriptBaseName}_${version}_win${bitness}.zip"
     echo zipName
 
     echo "Creating ZIP file"
     cd $distDir
     zip -r $zipName $scriptBaseName
-    cd $workDir
+    cd $workDir 
 
     echo "Deleting build directory"
     rm -r $workDir"/build"
-    rm -r  $distDir/$scriptBaseName
+    rm -r  "${distDir:?}/${scriptBaseName}"
 }
 
 echo "64 bit Python"
 
-if [ -d $HOME"/.wine/drive_c/Python27_64" ]; then  
+if [ -d '${HOME}/.wine/drive_c/Python27_64' ]; then
     echo "Python (64 bit) already installed"
 else
     echo "Python (64 bit) not yet installed, installing now"
     echo ""
-    installPython 64 $downloadURL64bit
+    installPython 64 $downloadURLP264bit
 fi
 
 # Get path to Python root
 pyRoot64=$(ls -d ~/.wine/drive_c/Python27_64/python-*)
 
 # Python interpreter
-python64=$pyRoot64"/python.exe" 
+python64="${pyRoot64}/python.exe"
 
 # Install PyInstaller (if not installed already)
 installPyInstaller $python64
 
 echo "32 bit Python"
 
-if [ -d $HOME"/.wine/drive_c/Python27_32" ]; then  
+if [ -d "$HOME/.wine/drive_c/Python27_32" ]; then
     echo "Python (32 bit) already installed"
 else
     echo "Python (32 bit) not yet installed, installing now"
     echo ""
-    installPython 32 $downloadURL32bit
+    installPython 32 $downloadURLP232bit
 fi
 
 # Get path to Python root
 pyRoot32=$(ls -d ~/.wine/drive_c/Python27_32/python-*)
 
 # Python interpreter
-python32=$pyRoot32"/python.exe" 
+python32="${pyRoot32}/python.exe"
 
 # Install PyInstaller (if not installed already)
 installPyInstaller $python32
@@ -135,4 +140,3 @@ buildBinaries 64 $pyRoot64 $python64 $specFile64bit
 
 echo "Building binaries, 32 bit"
 buildBinaries 32 $pyRoot32 $python32 $specFile32bit
-
